requirements.txt
openenv-core>=0.2.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.22.0
openai>=1.0.0
